requirements.txt
celery
redis
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
django-filter
psycopg2-binary
python-dotenv
django-environ
django-rest-auth
django-allauth
django-storages
boto3
gunicorn
whitenoise
Pillow